Maybe fix

+++ common/src/main/java/earth/terrarium/botarium/api/fluid/FluidHooks.java
@@ -46,8 +46,8 @@ public class FluidHooks {
      * @return The number of fluid values in the given number of buckets.
      */
     @ImplementedByExtension
-    public static long buckets(float buckets) {
-        return (long) buckets * getBucketAmount();
+    public static long buckets(double buckets) {
+        return (long) (buckets * getBucketAmount());
     }
 
     /**
